Moved business logic from NeighbourDiscoverer to NeighbourService
