+++ neighbour-service/src/main/java/no/vegvesen/ixn/federation/discoverer/NeighbourDiscovererProperties.java
