+++ neighbour-service/src/main/java/no/vegvesen/ixn/federation/service/NeighbourService.java
@@ -6,6 +6,9 @@ import no.vegvesen.ixn.federation.api.v1_0.SubscriptionRequestApi;
 import no.vegvesen.ixn.federation.api.v1_0.SubscriptionStatus;
 import no.vegvesen.ixn.federation.capability.JMSSelectorFilterFactory;
 import no.vegvesen.ixn.federation.discoverer.DNSFacade;
+import no.vegvesen.ixn.federation.discoverer.GracefulBackoffProperties;
+import no.vegvesen.ixn.federation.discoverer.NeighbourDiscovererProperties;
+import no.vegvesen.ixn.federation.discoverer.facade.NeighbourRESTFacade;
 import no.vegvesen.ixn.federation.exceptions.*;
 import no.vegvesen.ixn.federation.model.*;
 import no.vegvesen.ixn.federation.repository.NeighbourRepository;
@@ -21,13 +24,13 @@ import org.springframework.stereotype.Component;
 
 import java.time.LocalDateTime;
 import java.util.Collections;
+import java.util.List;
 import java.util.Set;
 
 @Component
 public class NeighbourService {
 	private static Logger logger = LoggerFactory.getLogger(NeighbourService.class);
 
-	@Value("${interchange.node-provider.name}")
 	private String myName;
 
 	private NeighbourRepository neighbourRepository;
@@ -37,12 +40,25 @@ public class NeighbourService {
 	private SubscriptionRequestTransformer subscriptionRequestTransformer = new SubscriptionRequestTransformer(subscriptionTransformer);
 	private DNSFacade dnsFacade;
 
+	private GracefulBackoffProperties backoffProperties;
+	private NeighbourDiscovererProperties discovererProperties;
+	private NeighbourRESTFacade neighbourRESTFacade;
+
+
 	public NeighbourService(NeighbourRepository neighbourRepository,
 							SelfRepository selfRepository,
-							DNSFacade dnsFacade) {
+							DNSFacade dnsFacade,
+							GracefulBackoffProperties backoffProperties,
+							NeighbourDiscovererProperties discovererProperties,
+							NeighbourRESTFacade neighbourRESTFacade,
+							@Value("${interchange.node-provider.name}") String myName) {
 		this.neighbourRepository = neighbourRepository;
 		this.selfRepository = selfRepository;
 		this.dnsFacade = dnsFacade;
+		this.backoffProperties = backoffProperties;
+		this.discovererProperties = discovererProperties;
+		this.neighbourRESTFacade = neighbourRESTFacade;
+		this.myName = myName;
 	}
 
 	public Self getSelf() {
@@ -184,4 +200,165 @@ public class NeighbourService {
 										dnsFacade.getDnsServerName()))
 				);
 	}
+
+	public List<Neighbour> findNeighboursToExhangeCapabilities() {
+		return neighbourRepository.findByCapabilities_StatusIn(
+				Capabilities.CapabilitiesStatus.UNKNOWN,
+				Capabilities.CapabilitiesStatus.KNOWN,
+				Capabilities.CapabilitiesStatus.FAILED);
+	}
+
+
+	public List<Neighbour> findNeighboursWithKnownCapabilities() {
+		return neighbourRepository.findByCapabilities_Status(Capabilities.CapabilitiesStatus.KNOWN);
+	}
+
+	public List<Neighbour> getNeighboursFailedSubscriptionRequest() {
+		return neighbourRepository.findByFedIn_StatusIn(SubscriptionRequestStatus.FAILED);
+	}
+
+	public void checkForNewNeighbours() {
+		logger.info("Checking DNS for new neighbours using {}.", dnsFacade.getClass().getSimpleName());
+		List<Neighbour> neighbours = dnsFacade.getNeighbours();
+		logger.debug("Got neighbours from DNS {}.", neighbours);
+
+		for (Neighbour neighbour : neighbours) {
+			NeighbourMDCUtil.setLogVariables(myName, neighbour.getName());
+			if (neighbourRepository.findByName(neighbour.getName()) == null && !neighbour.getName().equals(myName)) {
+
+				// Found a new Neighbour. Set capabilities status of neighbour to UNKNOWN to trigger capabilities exchange.
+				neighbour.getCapabilities().setStatus(Capabilities.CapabilitiesStatus.UNKNOWN);
+				logger.info("Found a new neighbour. Saving in database");
+				neighbourRepository.save(neighbour);
+			}
+			NeighbourMDCUtil.removeLogVariables();
+		}
+	}
+
+	public void capabilityExchange(List<Neighbour> neighboursForCapabilityExchange) {
+		Self self = selfRepository.findByName(myName);
+		if (self == null) {
+			logger.info("No representation of self. Skipping.");
+			return;
+		}
+
+		for (Neighbour neighbour : neighboursForCapabilityExchange) {
+			NeighbourMDCUtil.setLogVariables(myName, neighbour.getName());
+			logger.info("Posting capabilities to neighbour: {} ", neighbour.getName());
+			try {
+				if (backoffProperties.canBeContacted(neighbour)) {
+					if (neighbour.needsOurUpdatedCapabilities(self.getLastUpdatedLocalCapabilities())) {
+						Capabilities capabilities = neighbourRESTFacade.postCapabilitiesToCapabilities(self, neighbour);
+						capabilities.setLastCapabilityExchange(LocalDateTime.now());
+						neighbour.setCapabilities(capabilities);
+						neighbour.okConnection();
+						logger.info("Successfully completed capability exchange.");
+						logger.debug("Updated neighbour: {}", neighbour.toString());
+					} else {
+						logger.debug("Neighbour has our last capabilities");
+					}
+				} else {
+					logger.info("Too soon to post capabilities to neighbour when backing off");
+				}
+			} catch (CapabilityPostException e) {
+				neighbour.getCapabilities().setStatus(Capabilities.CapabilitiesStatus.FAILED);
+				neighbour.failedConnection(backoffProperties.getNumberOfAttempts());
+			} finally {
+				neighbour = neighbourRepository.save(neighbour);
+				logger.info("Saving updated neighbour: {}", neighbour.toString());
+				NeighbourMDCUtil.removeLogVariables();
+			}
+		}
+
+	}
+
+	public void evaluateAndPostSubscriptionRequest(List<Neighbour> neighboursForSubscriptionRequest) {
+		Self self = selfRepository.findByName(myName);
+		if (self == null) {
+			logger.warn("No local capabilities nor subscriptions.");
+			return; // We have nothing to post to our neighbour
+		}
+		LocalDateTime lastUpdatedLocalSubscriptions = self.getLastUpdatedLocalSubscriptions();
+
+		for (Neighbour neighbour : neighboursForSubscriptionRequest) {
+			NeighbourMDCUtil.setLogVariables(myName, neighbour.getName());
+			if (neighbour.shouldCheckSubscriptionRequestsForUpdates(lastUpdatedLocalSubscriptions)) {
+				try {
+					if (neighbour.hasEstablishedSubscriptions() || neighbour.hasCapabilities()) {
+						logger.info("Found neighbour for subscription request: {}", neighbour.getName());
+						Set<Subscription> calculatedSubscriptionForNeighbour = self.calculateCustomSubscriptionForNeighbour(neighbour);
+						Set<Subscription> fedInSubscriptions = neighbour.getFedIn().getSubscriptions();
+						if (!calculatedSubscriptionForNeighbour.equals(fedInSubscriptions)) {
+							try {
+								if (backoffProperties.canBeContacted(neighbour)) {
+									SubscriptionRequest subscriptionRequestResponse = neighbourRESTFacade.postSubscriptionRequest(self, neighbour, calculatedSubscriptionForNeighbour);
+									subscriptionRequestResponse.setSuccessfulRequest(LocalDateTime.now());
+									neighbour.setFedIn(subscriptionRequestResponse);
+									neighbour.okConnection();
+									logger.info("Successfully posted subscription request to neighbour.");
+								} else {
+									logger.info("Too soon to post subscription request to neighbour when backing off");
+								}
+							} catch (SubscriptionRequestException e) {
+								neighbour.getFedIn().setStatus(SubscriptionRequestStatus.FAILED);
+								neighbour.failedConnection(backoffProperties.getNumberOfAttempts());
+								logger.error("Failed subscription request. Setting status of neighbour fedIn to FAILED.\n", e);
+							}
+							// Successful subscription request, update discovery state subscription request timestamp.
+							neighbour = neighbourRepository.save(neighbour);
+							logger.info("Saving updated neighbour: {}", neighbour.toString());
+						}
+					}
+				} catch (Exception e) {
+					logger.error("Exception when evaluating subscriptions for neighbour", e);
+				}
+			}
+			NeighbourMDCUtil.removeLogVariables();
+		}
+	}
+
+	public void pollSubscriptions() {
+		List<Neighbour> neighboursToPoll = neighbourRepository.findNeighboursByFedIn_Subscription_SubscriptionStatusIn(
+				SubscriptionStatus.REQUESTED,
+				SubscriptionStatus.ACCEPTED,
+				SubscriptionStatus.FAILED);
+		for (Neighbour neighbour : neighboursToPoll) {
+			if (backoffProperties.canBeContacted(neighbour)) {
+				pollSubscriptionsOneNeighbour(neighbour, neighbour.getSubscriptionsForPolling());
+			}
+		}
+	}
+
+	public void pollSubscriptionsOneNeighbour(Neighbour neighbour, Set<Subscription> subscriptions) {
+		try {
+			NeighbourMDCUtil.setLogVariables(myName, neighbour.getName());
+			for (Subscription subscription : subscriptions) {
+				try {
+					if (subscription.getNumberOfPolls() < discovererProperties.getSubscriptionPollingNumberOfAttempts()) {
+						logger.info("Polling neighbour {} for status on subscription with path {}", neighbour.getName(), subscription.getPath());
+
+						// Throws SubscriptionPollException if unsuccessful
+						Subscription polledSubscription = neighbourRESTFacade.pollSubscriptionStatus(subscription, neighbour);
+						subscription.setSubscriptionStatus(polledSubscription.getSubscriptionStatus());
+						subscription.setNumberOfPolls(subscription.getNumberOfPolls() + 1);
+						neighbour.okConnection();
+						logger.info("Successfully polled subscription. Subscription status: {}  - Number of polls: {}", subscription.getSubscriptionStatus(), subscription.getNumberOfPolls());
+					} else {
+						// Number of poll attempts exceeds allowed number of poll attempts.
+						subscription.setSubscriptionStatus(SubscriptionStatus.GIVE_UP);
+						logger.warn("Number of polls has exceeded number of allowed polls. Setting subscription status to GIVE_UP.");
+					}
+				} catch (SubscriptionPollException e) {
+					subscription.setSubscriptionStatus(SubscriptionStatus.FAILED);
+					neighbour.failedConnection(backoffProperties.getNumberOfAttempts());
+					logger.error("Error in polling for subscription status. Setting status of Subscription to FAILED.");
+				}
+			}
+		} finally {
+			logger.info("Saving updated neighbour: {}", neighbour.toString());
+			neighbourRepository.save(neighbour);
+			NeighbourMDCUtil.removeLogVariables();
+		}
+	}
+
 }
