+++ neighbour-service/src/test/java/no/vegvesen/ixn/federation/service/NeighbourServiceTest.java
@@ -2,6 +2,9 @@ package no.vegvesen.ixn.federation.service;
 
 import no.vegvesen.ixn.federation.api.v1_0.*;
 import no.vegvesen.ixn.federation.discoverer.DNSFacade;
+import no.vegvesen.ixn.federation.discoverer.GracefulBackoffProperties;
+import no.vegvesen.ixn.federation.discoverer.NeighbourDiscovererProperties;
+import no.vegvesen.ixn.federation.discoverer.facade.NeighbourRESTFacade;
 import no.vegvesen.ixn.federation.exceptions.InterchangeNotInDNSException;
 import no.vegvesen.ixn.federation.exceptions.SubscriptionRequestException;
 import no.vegvesen.ixn.federation.model.*;
@@ -31,12 +34,19 @@ class NeighbourServiceTest {
 	DNSFacade dnsFacade;
 	@Mock
 	SelfRepository selfRepository;
+	@Mock
+	NeighbourRESTFacade neighbourRESTFacade;
+
+
+	private NeighbourDiscovererProperties discovererProperties = new NeighbourDiscovererProperties();
+	private GracefulBackoffProperties backoffProperties = new GracefulBackoffProperties();
+	private String myName = "bouvet.itsinterchange.eu";
 
 	NeighbourService neighbourService;
 
 	@BeforeEach
 	void setUp() {
-		neighbourService = new NeighbourService(neighbourRepository, selfRepository, dnsFacade);
+		neighbourService = new NeighbourService(neighbourRepository, selfRepository, dnsFacade, backoffProperties, discovererProperties, neighbourRESTFacade, myName);
 	}
 
 	@Test
